+++ src/components/Hidden/tests/Hidden.test.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { render, screen } from "@testing-library/react";
 import Hidden from "components/Hidden";
 
@@ -9,12 +8,6 @@ const fixtures = {
 describe("Utilities/Hidden", () => {
 	test("renders children", () => {
 		render(<Hidden>{fixtures.content}</Hidden>);
-
-		expect(screen.getByText(fixtures.content)).toBeInTheDocument();
-	});
-
-	test("works with render props", () => {
-		render(<Hidden>{(className) => <div className={className}>{fixtures.content}</div>}</Hidden>);
 
 		expect(screen.getByText(fixtures.content)).toBeInTheDocument();
 	});